% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{find.nested.blocks}
\alias{find.nested.blocks}
\title{Find nested blocks in a text}
\usage{
find.nested.blocks(txt, open = "#<", close = "#>")
}
\arguments{
\item{txt}{a character vector}

\item{open}{start of a block's header line, by default "#<"}

\item{close}{start of a block's trailer line by default "#>".}
}
\description{
Blocks must start and end with separate header
and trailer lines, respectively e.g.
}
\details{
#< collapse ", "

#>
}
