% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xglue.R
\name{xglue.pre}
\alias{xglue.pre}
\title{Just compile pre blocks and return resulting text}
\usage{
xglue.pre(
  txt,
  envir = parent.frame(),
  open = "<<",
  close = ">>",
  enclos = parent.frame(),
  newline = "<<newline>>",
  bdf = NULL
)
}
\arguments{
\item{txt}{The template text on which xglue operations shall be performed}

\item{envir}{Environment or list that contains objects whose value is spliced in. By default the environment from which xglue is called.}

\item{open}{the opening string of the to be replaced glue whiskers in pre block}

\item{close}{the closing string of the to be replaced glue whiskers in pre block}

\item{enclos}{If envir is a list the enclosing environment.}
}
\value{
The resulting text of the template were pre blocks have been replaced
}
\description{
Just compile pre blocks and return resulting text
}
