% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xglue.R
\name{xglue}
\alias{xglue}
\title{Perform xglue operation on a template text}
\usage{
xglue(
  txt,
  envir = parent.frame(),
  open = "{",
  close = "}",
  enclos = parent.frame(),
  bdf = parse.xglue.blocks(txt)
)
}
\arguments{
\item{txt}{The template text on which xglue operations shall be performed}

\item{envir}{Environment or list that contains objects whose value is spliced in. By default the environment from which xglue is called.}

\item{open}{the opening string of glue whiskers}

\item{close}{the closing string of glue whiskers}

\item{enclos}{If envir is a list the enclosing environment.}

\item{bdf}{the result of a manual call of parse.xglue.blocks. Don't need to change this parameter, may only be helpful for speeding things up if the same txt shall be glued with different with values.}
}
\value{
The glued text as a single character
}
\description{
Perform xglue operation on a template text
}
