% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xglue.R
\name{xglue}
\alias{xglue}
\title{Perform xglue operation on a template text}
\usage{
xglue(
  txt,
  envir = parent.frame(),
  open = "{",
  close = "}",
  pre.open = "<<",
  pre.close = ">>",
  block.open = "#<",
  block.close = "#>",
  enclos = parent.frame(),
  newline = "<<newline>>",
  just.pre = FALSE
)
}
\arguments{
\item{txt}{The template text on which xglue operations shall be performed}

\item{envir}{Environment or list that contains objects whose value is spliced in. By default the environment from which xglue is called.}

\item{open}{the opening string of glue whiskers}

\item{close}{the closing string of glue whiskers}

\item{pre.open}{the opening string of glue whiskers when pre compiling pre blocks}

\item{pre.close}{the closing string of glue whiskers when pre compiling pre blocks}

\item{block.open}{start of a block's header line, by default "#<"}

\item{block.close}{start of a block's trailer line by default "#>".}

\item{enclos}{If envir is a list the enclosing environment.}

\item{just.pre}{Just compile the pre blocks and return the resulting template.}
}
\value{
The glued text as a single character
}
\description{
Perform xglue operation on a template text
}
